Refactor bridge integration tests
