+++ integrationTests/chainSimulator/sovereignBridge/testdata/chain-config.wasm
