+++ integrationTests/chainSimulator/sovereignBridge/testdata/chain-factory.wasm
