+++ integrationTests/chainSimulator/sovereignBridge/testdata/mvx-esdt-safe.wasm
