+++ integrationTests/chainSimulator/sovereignBridge/testdata/mvx-fee-market.wasm
