+++ integrationTests/chainSimulator/sovereignBridge/testdata/sovereign-forge.wasm
